requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
enum34==1.1.6
fabric==2.1.3
funcsigs==1.0.2
idna==2.6
invoke==1.0.0
ipaddress==1.0.22
mock==2.0.0
paramiko==2.4.1
pbr==4.0.3
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
